# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Taleed Talent Application API
  version: 1.0.0
  description: Local Phase 2 identity and session contract. Business domain endpoints are added in later phases.
servers:
  - url: https://talent.taleed.test:9443
security:
  - laravelSession: []
paths:
...
  /api/v1/admin/organizations:
    post:
      operationId: createOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationRequest'
      responses:
        '201':
          description: Organization created.
        '403':
          $ref: '#/components/responses/Forbidden'
        '419':
          $ref: '#/components/responses/CsrfExpired'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    laravelSession:
      type: apiKey
      in: cookie
      name: taleed-talent-session
...
  schemas:
...
    OrganizationRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 255
        sector:
          type: string
        city:
          type: string
        timezone:
          type: string
...
  responses:
...
    Forbidden:
      description: The authenticated application user is not authorized.
    CsrfExpired:
      description: CSRF token is missing or expired.
    ValidationError:
      description: Request validation failed.
...